requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.0
supabase==2.7.4
aiohttp==3.9.1
asyncio-throttle==1.0.2
Pillow==10.2.0
PyPDF2==3.0.1